Fixed CustomData
If you use SAM with OFP then SAM allows you to put configuration data in the cockpit customdata.  The opening tag for SAM is SAM.  which isn't able to be parsed by the ingame engine using the INI parser.  If an error is found currently OFP will use a regex to look for all OFP settings instead of the ingame ini parser.